requirements.txt
asgiref==3.10.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
django-phonenumber-field==8.3.0
django-phonenumbers==1.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.43.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
phonenumbers==9.0.18
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
resend==2.19.0
rest-framework-simplejwt==0.0.2
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
